Resizing images, looping through csv

+++ poster.py
@@ -8,7 +8,7 @@ load_dotenv()
 username = os.getenv('USERNAMEIG')
 password = os.getenv('PASSWORD')
 
-shutil.rmtree('config')
+shutil.rmtree('config') if os.path.exists('config') else None
 
 bot = Bot()
 bot.login(username=username, password=password)
@@ -20,9 +20,13 @@ with open('products.csv', encoding='utf-8', newline='') as csv_file:
         images = row[3].split('|')
         caption = row[2]
 
-    for image in images:
-        img = Image.open('images/' + image)
-        img.resize((1080, 1080))
-        img.save('images/' + image)
+        for image in images:
+            img = Image.open('images/' + image)
+            img.resize((1080, 1080))
+            img.save('images/' + image, 'JPEG', quality=100)
 
-        bot.upload_photo('images/' + image, caption)
+    #     bot.upload_photo('images/' + image, caption)
+
+        with open('posted.csv', 'a', encoding='utf-8', newline='') as csvf:
+            writer = csv.writer(csvf)
+            writer.writerow([row[0]])
